requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.9.1
aiosignal==1.3.1
alabaster==1.0.0
amqp==5.2.0
asgiref==3.8.1
astroid==3.3.3
attrs==24.2.0
babel==2.16.0
billiard==4.2.1
celery==5.4.0
certifi==2024.8.30
cffi==1.17.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cron-descriptor==1.4.5
cryptography==43.0.0
dj-rest-auth==6.0.0
Django==5.1
django-allauth==64.2.0
django-celery-beat==2.7.0
django-cors-headers==4.4.0
django-extensions==3.2.3
django-timezone-field==7.0
djangorestframework==3.15.2
dnspython==2.6.1
docutils==0.21.2
drf-yasg==1.21.7
et-xmlfile==1.1.0
flower==2.0.1
frozenlist==1.5.0
humanize==4.10.0
idna==3.8
imagesize==1.4.1
inflection==0.5.1
Jinja2==3.1.4
jwt==1.3.1
kombu==5.4.2
MarkupSafe==2.1.5
multidict==6.1.0
numpy==2.1.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.3
pillow==10.4.0
prometheus_client==0.21.0
prompt_toolkit==3.0.47
propcache==0.2.0
pycparser==2.22
Pygments==2.18.0
PyJWT==2.9.0
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
redis==5.0.8
requests==2.32.3
setuptools==74.0.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==8.0.2
sphinx-autoapi==3.3.1
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
sqlparse==0.5.1
tornado==6.4.1
twilio==9.3.6
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
vine==5.1.0
wcwidth==0.2.13
yarl==1.17.1
